backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
pymongo==4.9.2
motor==3.6.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
celery==5.4.0
redis==5.2.1
google-generativeai==0.8.3
PyGithub==2.5.0
python-socketio>=5.11.0
python-engineio>=4.8.0
